CvSubdiv2DEdge is long
git-svn-id: https://emgucv.svn.sourceforge.net/svnroot/emgucv/trunk@130 d7f09016-e345-0410-b530-edf29a71df78

+++ Emgu.CV/PInvoke/CvType/MCvSubdiv2DPoint.cs
@@ -9,7 +9,7 @@ namespace Emgu.CV
     public struct MCvSubdiv2DPoint
     {
         public int flags;
-        public IntPtr first;
+        public long first;
         public MCvPoint2D32f pt;
     }
 }
